+++ debug_dem.cpp
@@ -11,11 +11,11 @@ using json = nlohmann::json;
 // Nodata = -999999
 
 int main() {
-    union {float value; uint8_t bytes[sizeof(float)];} t{};
+    union {double value; uint8_t bytes[sizeof(double)];} t{};
 
-    auto serialize = [&t](float value) -> float {
+    auto serialize = [&t](double value) -> double {
         t.value = value;
-        std::reverse(t.bytes, t.bytes + sizeof(float));
+        std::reverse(t.bytes, t.bytes + sizeof(double));
         return t.value;
     };
 	ifstream f("./metadata.json");
@@ -28,16 +28,16 @@ int main() {
 			cout << key << '\n';
 			cout << "DEM BIN PATH: " << value["DEM"]["bin_path"] << '\n';
             string bin_path = value["DEM"]["bin_path"];
-			float dem_x_min = value["DEM"]["x"]["min"];
-			float dem_x_max = value["DEM"]["x"]["max"];
-			float dem_y_min = value["DEM"]["y"]["min"];
-			float dem_y_max = value["DEM"]["y"]["max"];
+			double dem_x_min = value["DEM"]["x"]["min"];
+			double dem_x_max = value["DEM"]["x"]["max"];
+			double dem_y_min = value["DEM"]["y"]["min"];
+			double dem_y_max = value["DEM"]["y"]["max"];
 			cout << "DEM X MIN AND MAX: " << dem_x_min << " " << dem_x_max << '\n';
 			cout << "DEM Y MIN AND MAX: " << dem_y_min << " " << dem_y_max << '\n';
 			ifstream bin_file(bin_path, ios::binary);
             for (int i = 0; i < 10; i++) {
-                float t_value;
-                bin_file.read(reinterpret_cast<char*>(&t_value), sizeof(float));
+                double t_value;
+                bin_file.read(reinterpret_cast<char*>(&t_value), sizeof(double));
                 t_value = serialize(t_value);
                 cout << t_value << '\n';
             }
